corrected mistakes on asserts
