+++ src/Controller/LessonController.php
@@ -5,7 +5,7 @@ namespace App\Controller;
 use App\Entity\User;
 use App\Entity\Lesson;
 use App\Entity\Tutorial;
-use App\Form\LessonType;
+use App\Form\QuizLessonType;
 use App\Entity\Explanation;
 use App\Repository\LessonRepository;
 use Symfony\Component\HttpFoundation\Request;
@@ -37,7 +37,7 @@ class LessonController extends AbstractController
         $tutorial = $lesson->getTutorial();
         $quizzDone = $lesson->getUsers()->contains($this->getUser());
         if (!$quizzDone) {
-            $form = $this->createForm(LessonType::class, $lesson);
+            $form = $this->createForm(QuizLessonType::class, $lesson);
             $form->handleRequest($request);
             if ($form->isSubmitted() && $form->isValid()) {
                 /** @var User */
